+++ internal/deploy/edge.go
@@ -35,7 +35,7 @@ func DeployEdge(o EdgeOpts) error {
 		o.PrimaryUser = "root"
 	}
 	if o.Version == "" {
-		o.Version = "0.8.3"
+		o.Version = "0.8.4"
 	}
 	if o.AgentArch == "" {
 		o.AgentArch = "arm64"